+++ packages/vibrant-components/src/lib/Table/Table.tsx
@@ -1,8 +1,10 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+/* eslint-disable max-lines */
 /* eslint-disable @typescript-eslint/naming-convention */
 import type { ReactElement } from 'react';
-import { Children, isValidElement, useEffect, useMemo, useState } from 'react';
-import { Box, Image, ScrollBox, useConfig } from '@vibrant-ui/core';
-import { isDefined, useControllableState } from '@vibrant-ui/utils';
+import { Children, isValidElement, useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { Box, Image, ScrollBox, isNative, useConfig } from '@vibrant-ui/core';
+import { isDefined } from '@vibrant-ui/utils';
 import { Body } from '../Body';
 import { GhostButton } from '../GhostButton';
 import { HStack } from '../HStack';
@@ -12,12 +14,82 @@ import { VStack } from '../VStack';
 import { TableColumn } from './TableColumn';
 import type { TableColumnProps } from './TableColumn/TableColumnProps';
 import { TableDataCell } from './TableDataCell';
+import type { TableDataCellProps } from './TableDataCell/TableDataCellProps';
 import { TableHeaderCell } from './TableHeaderCell';
-import type { TableProps, TableSortBy, UseTableResult } from './TableProps';
+import type { TableHeaderCellProps } from './TableHeaderCell/TableHeaderCellProps';
+import type { TableCellRange, TableProps, TableSortBy, UseTableResult } from './TableProps';
 import { TableRow } from './TableRow';
 import type { SortDirection } from './TableSortIcon';
 
 const getCellKey = (key: any, rowIndex: number) => `${key}:${rowIndex}`;
+
+const getCornersFromSelectedRange = (selectedRange: TableCellRange) => {
+  const { anchor, cursor } = selectedRange;
+
+  return {
+    startRow: Math.min(anchor.rowIdx, cursor.rowIdx),
+    endRow: Math.max(anchor.rowIdx, cursor.rowIdx),
+    startCol: Math.min(anchor.colIdx, cursor.colIdx),
+    endCol: Math.max(anchor.colIdx, cursor.colIdx),
+  };
+};
+
+const isCellInSelectedRange = (
+  rowIdx: number,
+  colIdx: number,
+  multiCellSelectable: boolean,
+  selectedRange?: TableCellRange
+) => {
+  if (!multiCellSelectable || !selectedRange) {
+    return false;
+  }
+
+  const { startRow, endRow, startCol, endCol } = getCornersFromSelectedRange(selectedRange);
+
+  return rowIdx >= startRow && rowIdx <= endRow && colIdx >= startCol && colIdx <= endCol;
+};
+
+const isCellOnEdgeOfSelectedRange = (
+  rowIdx: number,
+  colIdx: number,
+  multiCellSelectable: boolean,
+  selectingRangeFromCell: boolean,
+  selectedRange?: TableCellRange
+) => {
+  const cellOnEdge: TableDataCellProps['selectedOnEdge'] = { top: true, bottom: true, left: true, right: true };
+
+  if (!multiCellSelectable || !selectedRange) {
+    return cellOnEdge;
+  }
+
+  const { startRow, endRow, startCol, endCol } = getCornersFromSelectedRange(selectedRange);
+
+  cellOnEdge.top = rowIdx === startRow && colIdx >= startCol && colIdx <= endCol && selectingRangeFromCell;
+  cellOnEdge.bottom = rowIdx === endRow && colIdx >= startCol && colIdx <= endCol;
+  cellOnEdge.left = colIdx === startCol && rowIdx >= startRow && rowIdx <= endRow;
+  cellOnEdge.right = colIdx === endCol && rowIdx >= startRow && rowIdx <= endRow;
+
+  return cellOnEdge;
+};
+
+const isHeaderOnEdgeOfSelectedRange = (
+  colIdx: number,
+  multiCellSelectable: boolean,
+  selectedRange?: TableCellRange
+) => {
+  const cellOnEdge: TableHeaderCellProps['selectedOnEdge'] = { left: true, right: true };
+
+  if (!multiCellSelectable || !selectedRange) {
+    return cellOnEdge;
+  }
+
+  const { startCol, endCol } = getCornersFromSelectedRange(selectedRange);
+
+  cellOnEdge.left = colIdx === startCol;
+  cellOnEdge.right = colIdx === endCol;
+
+  return cellOnEdge;
+};
 
 export const useTable = <Data extends Record<string, any>, RowKey extends keyof Data>(): UseTableResult<
   Data,
@@ -31,6 +103,7 @@ export const Table = <Data extends Record<string, any>, RowKey extends keyof Dat
   rowKey,
   loading = false,
   selectable = false,
+  multiCellSelectable = true,
   selectButtons,
   onSelectionChange,
   renderExpanded,
@@ -54,20 +127,26 @@ export const Table = <Data extends Record<string, any>, RowKey extends keyof Dat
     [children]
   );
 
-  const [sortBy, setSortBy] = useState<TableSortBy<Data>>({
-    direction: 'none',
-  });
+  const [sortBy, setSortBy] = useState<TableSortBy<Data>>();
 
-  useEffect(() => {
-    const defaultSort = columns.find(column => column.sortDirection && column.sortDirection !== 'none');
+  const defaultSortColumn = useMemo(
+    () => columns.find(column => column.sortDirection && column.sortDirection !== 'none'),
+    [columns]
+  );
 
-    if (defaultSort && defaultSort.dataKey !== sortBy.dataKey) {
-      setSortBy({
-        dataKey: defaultSort.dataKey,
-        direction: defaultSort.sortDirection ?? 'none',
-      });
-    }
-  }, [columns, sortBy.dataKey]);
+  const defaultSortBy = useMemo(
+    () => ({
+      dataKey: defaultSortColumn?.dataKey,
+      direction: defaultSortColumn?.sortDirection ?? 'none',
+    }),
+    [defaultSortColumn]
+  );
+
+  // sortBy가 존재하고 columns에 해당 dataKey가 있을 때만 sortBy 적용 (columns 변경에 대비)
+  const finalSortBy = useMemo(
+    () => (sortBy && columns.some(column => column.dataKey === sortBy.dataKey) ? sortBy : defaultSortBy),
+    [columns, defaultSortBy, sortBy]
+  );
 
   const handleChangeSort = (sortBy: TableSortBy<Data>) => {
     setSortBy(sortBy);
@@ -75,42 +154,158 @@ export const Table = <Data extends Record<string, any>, RowKey extends keyof Dat
     onSort?.(sortBy);
   };
 
-  const [selectedRowKeys, setSelectedRowKeys] = useControllableState<Set<Data[RowKey]>>({
-    defaultValue: new Set<Data[RowKey]>(),
-    onValueChange: (value: Set<Data[RowKey]>) => onSelectionChange?.([...value]),
-  });
+  const [selectedRows, setSelectedRows] = useState(new Set<Data>());
 
+  // 기존 onSelectionChange 동작을 보장하기 위해 rowKey로 호출할 수 있도록 변환
+  const mapRowsToRowKeys = useCallback((rows: Set<Data>) => [...Array(...rows).map(row => row[rowKey])], [rowKey]);
+
+  const handleSetSelectedRows = useCallback(
+    (value: Set<Data>) => {
+      setSelectedRows(value);
+      onSelectionChange?.(mapRowsToRowKeys(value));
+    },
+    [mapRowsToRowKeys, onSelectionChange]
+  );
+
+  // 데이터 변경 시에만 selectedRows 변경 및 onSelectionChange 호출하기 위해 사용
+  const prevData = useRef(data);
+
+  // data가 변경되면 data에 존재하지 않는 row를 selectedRows에서 제거
   useEffect(() => {
-    setSelectedRowKeys(new Set(data.filter(row => selectedRowKeys.has(row[rowKey])).map(row => row[rowKey])));
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [data, rowKey]);
+    if (prevData.current === data) {
+      return;
+    }
+
+    const dataSet = new Set(data);
+    // prevData에서 data에 존재하는 row를 제거하고 남은 row가 제거 대상
+    const toBeRemovedRows = new Set(prevData.current.filter(row => !dataSet.has(row)));
+
+    if (toBeRemovedRows.size > 0) {
+      setSelectedRows(prev => {
+        const newSet = new Set(prev);
+        const sizeBeforeRemove = newSet.size;
+
+        toBeRemovedRows.forEach(row => newSet.delete(row));
+
+        // 선택된 상태에서 제거된 row가 있을 때만 onSelectionChange 호출하고 상태를 변경한다
+        if (sizeBeforeRemove > newSet.size) {
+          onSelectionChange?.(mapRowsToRowKeys(newSet));
+
+          return newSet;
+        }
+
+        return prev;
+      });
+    }
+
+    prevData.current = data;
+  }, [data, mapRowsToRowKeys, onSelectionChange]);
 
   const [selectedCellKey, setSelectedCellKey] = useState<string>();
   const isCellClickEnabled = columns?.some(column => isDefined(column.onDataCell));
 
-  const handleToggleCheckbox = (key: Data[RowKey]) => {
-    const newSelectedRowKeys = new Set(selectedRowKeys);
+  const isSelectingRange = useRef(false);
+  const [selectedRange, setSelectedRange] = useState<TableCellRange>();
+  const [selectingRangeFromCell, setSelectingRangeFromCell] = useState(true);
 
-    if (newSelectedRowKeys.has(key)) {
-      newSelectedRowKeys.delete(key);
-    } else {
-      newSelectedRowKeys.add(key);
+  const tableRef = useRef<HTMLTableElement>(null);
+
+  // multiCellSelectable일 때, 테이블의 selectstart 이벤트를 막아서 텍스트 드래그 방지
+  useEffect(() => {
+    if (!tableRef.current) {
+      return;
     }
 
-    setSelectedRowKeys(newSelectedRowKeys);
+    const tableEl = tableRef.current;
+
+    const handleSelectStart = (event: Event) => {
+      if (multiCellSelectable) {
+        event.preventDefault();
+      }
+    };
+
+    tableEl.addEventListener('selectstart', handleSelectStart);
+
+    return () => {
+      tableEl.removeEventListener('selectstart', handleSelectStart);
+    };
+  }, [multiCellSelectable]);
+
+  const isShiftKeyPressed = useRef(false);
+
+  // Shift 키 누르고 있는 동안에만 range 선택 가능
+  useEffect(() => {
+    if (isNative) {
+      return;
+    }
+
+    const handleKeyDown = (event: KeyboardEvent) => {
+      if (event.key === 'Shift') {
+        isShiftKeyPressed.current = true;
+      }
+    };
+
+    const handleKeyUp = (event: KeyboardEvent) => {
+      if (event.key === 'Shift') {
+        isShiftKeyPressed.current = false;
+      }
+    };
+
+    window.addEventListener('keydown', handleKeyDown);
+    window.addEventListener('keyup', handleKeyUp);
+
+    return () => {
+      window.removeEventListener('keydown', handleKeyDown);
+      window.removeEventListener('keyup', handleKeyUp);
+    };
+  }, []);
+
+  const copySelectedCells = useCallback(() => {
+    if (!selectedRange) {
+      return;
+    }
+
+    const { startRow, endRow, startCol, endCol } = getCornersFromSelectedRange(selectedRange);
+
+    const selectedCells = [];
+
+    const columnNames = columns.slice(startCol, endCol + 1).map(column => column.title || column.dataKey || '');
+
+    selectedCells.push(columnNames.join('\t'));
+
+    for (let rowIdx = startRow; rowIdx <= endRow; rowIdx++) {
+      const row = data[rowIdx];
+      const rowCells = columns.slice(startCol, endCol + 1).map(column => {
+        const cellData = column.dataKey ? row[column.dataKey] : null;
+
+        return isDefined(column.formatData) ? column.formatData(row) : cellData;
+      });
+
+      selectedCells.push(rowCells.join('\t'));
+    }
+
+    const clipboardText = selectedCells.join('\n');
+
+    navigator?.clipboard.writeText(clipboardText);
+  }, [columns, data, selectedRange]);
+
+  const handleToggleCheckbox = (key: Data) => {
+    const newSelectedRows = new Set(selectedRows);
+
+    if (newSelectedRows.has(key)) {
+      newSelectedRows.delete(key);
+    } else {
+      newSelectedRows.add(key);
+    }
+
+    handleSetSelectedRows(newSelectedRows);
   };
 
   const handleToggleAllCheckbox = ({ value }: { value: boolean }) => {
     if (value) {
-      const allRowKeys = new Set<Data[RowKey]>(
-        data
-          .filter(row => (isDefined(disabledRowKeys) ? !disabledRowKeys.includes(row[rowKey]) : true))
-          .map(row => row[rowKey])
-      );
-
-      setSelectedRowKeys(allRowKeys);
+      handleSetSelectedRows(new Set(data.filter(row => !disabledRowKeys?.includes(row[rowKey]))));
     } else {
-      setSelectedRowKeys(new Set());
+      handleSetSelectedRows(new Set());
     }
   };
 
@@ -130,18 +325,20 @@ export const Table = <Data extends Record<string, any>, RowKey extends keyof Dat
       data-testid={testId}
     >
       <Box
+        ref={tableRef}
         as="table"
         display="web_table"
         borderCollapse="separate"
         tableLayout={tableLayout}
         width="100%"
         height="100%"
+        onCopy={copySelectedCells}
       >
         <Box as="thead" display="web_table-row-group" height="100%">
           <TableRow
             header={true}
             selectable={selectable}
-            selected={selectedRowKeys.size !== 0}
+            selected={selectedRows.size !== 0}
             expanded={!loading && data.length === 0}
             renderExpanded={() => (
               <VStack alignHorizontal="center" mt={32} mb={64}>
@@ -149,12 +346,12 @@ export const Table = <Data extends Record<string, any>, RowKey extends keyof Dat
                 <Body level={1}>{emptyText}</Body>
               </VStack>
             )}
-            overlaid={selectable && selectedRowKeys.size > 0}
+            overlaid={selectable && selectedRows.size > 0}
             renderOverlay={() => (
               <HStack alignVertical="center" height="100%" spacing={12}>
-                {selectedRowKeys.size > 0 ? (
+                {selectedRows.size > 0 ? (
                   <Body level={2} weight="medium">
-                    {tableTranslation.numberOfSelected.replace('{count}', selectedRowKeys.size.toString())}
+                    {tableTranslation.numberOfSelected.replace('{count}', selectedRows.size.toString())}
                   </Body>
                 ) : null}
                 <HStack alignVertical="center" height="100%" spacing={12}>
@@ -163,7 +360,7 @@ export const Table = <Data extends Record<string, any>, RowKey extends keyof Dat
                       key={text}
                       size="md"
                       color="onViewInformative"
-                      onClick={() => onClick(data.filter(row => selectedRowKeys.has(row[rowKey])))}
+                      onClick={() => onClick(data.filter(row => selectedRows.has(row[rowKey])))}
                     >
                       {text}
                     </GhostButton>
@@ -171,7 +368,7 @@ export const Table = <Data extends Record<string, any>, RowKey extends keyof Dat
                 </HStack>
               </HStack>
             )}
-            indeterminate={selectedRowKeys.size !== data.length}
+            indeterminate={selectedRows.size !== data.length}
             onSelectionChange={handleToggleAllCheckbox}
             disabled={loading}
           >
@@ -179,18 +376,21 @@ export const Table = <Data extends Record<string, any>, RowKey extends keyof Dat
               <TableHeaderCell renderCell={() => <Box width={16} height={16} />} width={48} />
             )}
             {columns.map(
-              ({
-                key,
-                dataKey,
-                alignHorizontal,
-                alignVertical,
-                renderHeader,
-                lineLimit,
-                wordBreak,
-                whiteSpace,
-                overflowWrap,
-                ...column
-              }: TableColumnProps<Data>) => (
+              (
+                {
+                  key,
+                  dataKey,
+                  alignHorizontal,
+                  alignVertical,
+                  renderHeader,
+                  lineLimit,
+                  wordBreak,
+                  whiteSpace,
+                  overflowWrap,
+                  ...column
+                }: TableColumnProps<Data>,
+                colIdx
+              ) => (
                 <TableHeaderCell
                   key={key}
                   {...column}
@@ -201,8 +401,37 @@ export const Table = <Data extends Record<string, any>, RowKey extends keyof Dat
                   whiteSpace={whiteSpace?.header}
                   overflowWrap={overflowWrap?.header}
                   renderCell={renderHeader}
-                  sortDirection={sortBy.dataKey === dataKey ? sortBy.direction : 'none'}
+                  sortDirection={finalSortBy.dataKey === dataKey ? finalSortBy.direction : 'none'}
                   onSort={(sortDirection: SortDirection) => handleChangeSort({ dataKey, direction: sortDirection })}
+                  multiCellSelectable={multiCellSelectable}
+                  onPressIn={() => {
+                    setSelectingRangeFromCell(false);
+                    isSelectingRange.current = true;
+                    setSelectedRange(prev => ({
+                      anchor:
+                        isShiftKeyPressed.current && !selectingRangeFromCell && prev
+                          ? { rowIdx: 0, colIdx: prev.anchor.colIdx }
+                          : { rowIdx: 0, colIdx },
+                      cursor: { rowIdx: data.length - 1, colIdx },
+                    }));
+                  }}
+                  onPressOut={() => {
+                    isSelectingRange.current = false;
+                  }}
+                  onHoverIn={() => {
+                    if (!isSelectingRange.current) {
+                      return;
+                    }
+
+                    setSelectedRange(prev => ({
+                      anchor: prev ? prev.anchor : { rowIdx: 0, colIdx },
+                      cursor: { rowIdx: selectingRangeFromCell ? 0 : data.length - 1, colIdx },
+                    }));
+                  }}
+                  selected={
+                    !selectingRangeFromCell && isCellInSelectedRange(0, colIdx, multiCellSelectable, selectedRange)
+                  }
+                  selectedOnEdge={isHeaderOnEdgeOfSelectedRange(colIdx, multiCellSelectable, selectedRange)}
                 />
               )
             )}
@@ -210,12 +439,12 @@ export const Table = <Data extends Record<string, any>, RowKey extends keyof Dat
         </Box>
         <Box as="tbody" display="web_table-row-group" height="100%">
           {!loading
-            ? data.map((row, index) => (
+            ? data.map((row, rowIdx) => (
                 <TableRow
-                  key={row[rowKey]}
+                  key={rowKey ? row[rowKey] : rowIdx}
                   selectable={selectable}
-                  selected={selectedRowKeys.has(row[rowKey])}
-                  onSelectionChange={() => handleToggleCheckbox(row[rowKey])}
+                  selected={selectedRows.has(row)}
+                  onSelectionChange={() => handleToggleCheckbox(row)}
                   expandable={isDefined(renderExpanded)}
                   renderExpanded={() => (
                     <Paper backgroundColor="surface1" p={16}>
@@ -226,38 +455,61 @@ export const Table = <Data extends Record<string, any>, RowKey extends keyof Dat
                   disabled={disabledRowKeys?.includes(row[rowKey])}
                 >
                   {columns.map(
-                    ({
-                      key,
-                      dataKey,
-                      alignHorizontal,
-                      alignVertical,
-                      lineLimit,
-                      wordBreak,
-                      whiteSpace,
-                      overflowWrap,
-                      onDataCell,
-                      formatData,
-                      renderDataCell,
-                      selectable: cellSelectable,
-                      width: _,
-                      ...column
-                    }: TableColumnProps<Data>) => (
+                    (
+                      {
+                        key,
+                        dataKey,
+                        alignHorizontal,
+                        alignVertical,
+                        lineLimit,
+                        wordBreak,
+                        whiteSpace,
+                        overflowWrap,
+                        onDataCell,
+                        formatData,
+                        renderDataCell,
+                        selectable: cellSelectable,
+                        width: _,
+                        ...column
+                      }: TableColumnProps<Data>,
+                      colIdx
+                    ) => (
                       <TableDataCell
                         key={key}
-                        onClick={
-                          isCellClickEnabled
-                            ? onDataCell
-                              ? () => {
-                                  onDataCell.onClick?.(row);
+                        onClick={() => {
+                          if (isCellClickEnabled && onDataCell) {
+                            onDataCell.onClick?.(row);
+                            setSelectedCellKey(getCellKey(key, rowIdx));
+                          } else if (onRow) {
+                            onRow.onClick?.(row);
+                          }
+                        }}
+                        onPressIn={() => {
+                          setSelectingRangeFromCell(true);
+                          isSelectingRange.current = true;
+                          setSelectedRange(prev => ({
+                            anchor: isShiftKeyPressed.current && prev ? prev.anchor : { rowIdx, colIdx },
+                            cursor: { rowIdx, colIdx },
+                          }));
+                        }}
+                        onPressOut={() => {
+                          isSelectingRange.current = false;
+                        }}
+                        onHoverIn={() => {
+                          if (!isSelectingRange.current) {
+                            return;
+                          }
 
-                                  setSelectedCellKey(getCellKey(key, index));
-                                }
-                              : undefined
-                            : onRow
-                            ? () => onRow.onClick?.(row)
-                            : undefined
-                        }
-                        onCopy={() => onDataCell?.onCopy?.(row)}
+                          setSelectedRange(prev => ({
+                            anchor: prev ? prev.anchor : { rowIdx, colIdx },
+                            cursor: { rowIdx: selectingRangeFromCell ? rowIdx : data.length - 1, colIdx },
+                          }));
+                        }}
+                        onCopy={() => {
+                          if (!multiCellSelectable) {
+                            onDataCell?.onCopy?.(row);
+                          }
+                        }}
                         alignVertical={alignVertical?.dataCell}
                         alignHorizontal={alignHorizontal?.dataCell}
                         lineLimit={lineLimit?.dataCell}
@@ -265,7 +517,18 @@ export const Table = <Data extends Record<string, any>, RowKey extends keyof Dat
                         whiteSpace={whiteSpace?.dataCell}
                         overflowWrap={overflowWrap?.dataCell}
                         disabled={disabledRowKeys?.includes(row[rowKey])}
-                        selected={cellSelectable && selectedCellKey === getCellKey(key, index)}
+                        selected={
+                          (cellSelectable && selectedCellKey === getCellKey(key, rowIdx)) ||
+                          (multiCellSelectable &&
+                            isCellInSelectedRange(rowIdx, colIdx, multiCellSelectable, selectedRange))
+                        }
+                        selectedOnEdge={isCellOnEdgeOfSelectedRange(
+                          rowIdx,
+                          colIdx,
+                          multiCellSelectable,
+                          selectingRangeFromCell,
+                          selectedRange
+                        )}
                         renderCell={renderDataCell ? () => renderDataCell?.(row) : undefined}
                         {...column}
                       >