+++ packages/vibrant-components/src/lib/Table/TableDataCell/TableDataCellProps.ts
@@ -11,11 +11,15 @@ export type TableDataCellProps = {
   children?: TextChildren;
   renderCell?: () => ReactElementChildren;
   onClick?: () => void;
+  onPressIn?: () => void;
+  onPressOut?: () => void;
+  onHoverIn?: () => void;
   onCopy?: () => void;
   alignVertical?: 'center' | 'end' | 'start';
   alignHorizontal?: 'center' | 'end' | 'start';
   disabled?: boolean;
   selected?: boolean;
+  selectedOnEdge?: { top: boolean; bottom: boolean; left: boolean; right: boolean };
 } & Pick<FlexboxSystemProps, 'flexBasis' | 'flexGrow' | 'flexShrink'> &
   Pick<SizingSystemProps, 'minWidth' | 'width'> &
   Pick<TextProps, 'lineLimit' | 'overflowWrap' | 'whiteSpace' | 'wordBreak'>;