+++ packages/vibrant-motion/src/lib/Motion/Motion.stories.tsx
@@ -20,9 +20,6 @@ export default {
       },
     },
   },
-  parameters: {
-    chromatic: { disableSnapshot: true },
-  },
 } as ComponentMeta<typeof Motion>;
 
 export const Basic: ComponentStory<typeof Motion> = props => (