+++ packages/utils-eslint-config/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.91.19](https://github.com/pedaling/opensource/compare/utils-eslint-config-0.91.18...utils-eslint-config-0.91.19) (2025-03-27)
+
+## [0.91.18](https://github.com/pedaling/opensource/compare/utils-eslint-config-0.91.17...utils-eslint-config-0.91.18) (2025-03-21)
+
+## [0.91.17](https://github.com/pedaling/opensource/compare/utils-eslint-config-0.91.16...utils-eslint-config-0.91.17) (2025-03-12)
+
 ## [0.91.16](https://github.com/pedaling/opensource/compare/utils-eslint-config-0.91.15...utils-eslint-config-0.91.16) (2025-03-04)
 
 ## [0.91.15](https://github.com/pedaling/opensource/compare/utils-eslint-config-0.91.14...utils-eslint-config-0.91.15) (2025-02-28)