+++ packages/workspace-plugin/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.73.19](https://github.com/pedaling/opensource/compare/workspace-plugin-0.73.18...workspace-plugin-0.73.19) (2025-03-27)
+
+## [0.73.18](https://github.com/pedaling/opensource/compare/workspace-plugin-0.73.17...workspace-plugin-0.73.18) (2025-03-21)
+
+## [0.73.17](https://github.com/pedaling/opensource/compare/workspace-plugin-0.73.16...workspace-plugin-0.73.17) (2025-03-12)
+
 ## [0.73.16](https://github.com/pedaling/opensource/compare/workspace-plugin-0.73.15...workspace-plugin-0.73.16) (2025-03-04)
 
 ## [0.73.15](https://github.com/pedaling/opensource/compare/workspace-plugin-0.73.14...workspace-plugin-0.73.15) (2025-02-28)